src/virtualhome/requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
matplotlib>=3.4.2
networkx==2.3
numpy>=1.21
opencv-python==4.5.1.48
pillow>=8.3.1
plotly==3.10.0
requests==2.21.0
ipdb==0.13.9
termcolor==1.1.0
tqdm==4.31.1
urllib3>=1.24.3
